requirements.txt
appdirs==1.4.4
argh==0.26.2
Flask-Bootstrap==3.3.7.1
flask-eureka==0.1.0
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.0.4
httplib2==0.20.2
httplib2shim==0.0.3
ifaddr==0.1.7
ipython==7.25.0
jedi==0.18.0
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
pathspec==0.7.0
PyMySQL==1.0.2
pyproj==3.3.0
pyreadline==2.1
pyshp==2.1.3
pytest==6.2.4
python-box==5.4.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
unicodecsv==0.14.1
websockets==10.1
WTForms==3.0.1
xlwt==1.3.0

